refactor motor naming/parameters
* rename left/right to m1/m2
* allow different gains, ticks for each motor

+++ include/etherbotix/etherbotix_motor.hpp
@@ -102,7 +102,7 @@ private:
 
   double velocity_;  // rad/s
   double position_;  // radians
-  float current_;    // amperes
+  double current_;   // amperes
 
   float kp_;
   float kd_;
